Solve euler problems 1-2. Fix iterate bug (partly).
IterateC (iterate's return type) works by growing when an iterator
reaches its end. Its iterators therefore are never equal, even if they
point to the same location. take_while constructed its return type via
S(begin,it), which causes an infinite loop and rightly so!

The solution was to have take_while call take, using the distance
between the iterators as n, but this is neither generic nor correct
since it uses subtraction, which only works for random access iterators,
of which IterateC's are not.

A better solution may be to define a class, like Functor or Monad,
called Dup that allows specialization of duplication behaviour.

Signed-off-by: Scott Prager <splinterofchaos@gmail.com>