Improve astar.
Instead of taking from the front or back of the fringe, expecting that
to be the minimal node, search to be sure.

With this improvement, the Manhattan heuristic outperforms the Hamming
and null heuristics. In fact, the null heuristic outperforms Hamming,
which shows that Hamming is actually a bad heuristic.

Signed-off-by: Scott Prager <splinterofchaos@gmail.com>